docs: documented cmd task
refactor: moved some widgets to their separated folder
