+++ Cargo.toml
@@ -12,9 +12,7 @@ tree-sitter = "0.24.4"
 tree-sitter-highlight = "0.24.4"
 tree-sitter-javascript = "0.23.1"
 unicode-normalization = "0.1.24"
-isht = { path = "./isht/" }
 gapbuf = "0.1.4"
 downcast-rs = "2.0.1"
 tachyonfx = "0.10.1"
-[workspace]
-members = ["./isht"]
+
