--- isht/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "isht"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = { version = "1.0.94", features = ["backtrace"] }
