--- isht/src/lib.rs
@@ -1,8 +1,0 @@
-pub mod configuration;
-mod lexer;
-mod parsing;
-mod tasks;
-pub use parsing::errors::IshtParseError;
-pub use parsing::parse_content;
-pub use parsing::statment::ConfigStatment;
-pub use tasks::CmdTask;
